refactor(LottoGame): 로또 게임에 구매와 추첨의 두 책임을 분리한다
