+++ src/main/java/lotto/domain/LottoGame.java
@@ -1,26 +1,13 @@
 package lotto.domain;
 
-import java.util.ArrayList;
-import java.util.List;
-
 public class LottoGame {
-    private final LottoMoney lottoMoney;
+    private final LottoTickets lottoTickets;
 
-    public LottoGame(LottoMoney lottoMoney) {
-        this.lottoMoney = lottoMoney;
+    public LottoGame(LottoTickets lottoTickets) {
+        this.lottoTickets = lottoTickets;
     }
 
-    public LottoTickets buy(LottoNumberStrategy lottoNumberStrategy) {
-        int availableCount = lottoMoney.getAvailableBuyingCount();
-        List<LottoTicket> lottoTickets = new ArrayList<>();
-        for (int i = 0; i < availableCount; i++) {
-            LottoTicket lottoTicket = new LottoTicket(lottoNumberStrategy.get());
-            lottoTickets.add(lottoTicket);
-        }
-        return new LottoTickets(lottoTickets);
-    }
-
-    public LottoGameResults match(LottoTickets lottoTickets, WinningTicket winningTicket) {
+    public LottoGameResults match(WinningTicket winningTicket) {
         return new LottoGameResults(lottoTickets.checkRank(winningTicket));
     }
 }
